Refactoring signup page
